service worker: spoof /changes request
Spoofs the /changes first changes request made to the page to trick the
client into thinking that there is changes, if there are queued changes
in the local database. Due to deployment constraints, the serviceworker
has a hard time altering the request to the sharebill index. This is a
workaround, such that every time when the page makes its initial
/changes request, if there are local changes that have not been
syncronized with the server, the client will be sent a fake response
indicating that there are changes available. This initiates the client's
"/recent" request, which is altered by the serviceworker to include any
queued transactions.